Refactor location-based notifications to use monitored regions instead
For some unknown reason, the use of UNLocationNotificationTrigger to
display a notification when entering a region was causing a conflict with
the Notification Content extension and custom notification UI, resulting
in it failing to display the associated list of shopping items.

Switching to explicit monitoring of regions using the CLLocationManager
and manually triggering a notification when entering a region seems to
solve the issue. However, the app requires permission to always access
the device location in order for region monitoring to function.