Update dependencies for compatibility with Swift 4.2
The latest version of AWS Mobile SDK does not have a framework binary
available for Swift 4.2 yet, so it has to be built from source instead.
This means that the build phase script to strip unused AWS frameworks
needs to be disabled until a binary version of the framework is available.